requirements.txt
annotated-types==0.7.0
anyio==4.8.0
beautifulsoup4>=4.12.0
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
distro==1.9.0
flask==3.0.2
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.8.2
MarkupSafe==3.0.2
openai==1.74.0
pydantic==2.10.6
pydantic_core==2.27.2
requests>=2.31.0
sniffio==1.3.1
soupsieve==2.6
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
gunicorn==21.2.0
python-dotenv==0.21.0
# gevent==23.9.1
psycopg2-binary==2.9.9
Flask-SQLAlchemy==3.1.1
boto3==1.34.34
SQLAlchemy==2.0.25
fake-useragent>=1.4.0
cloudscraper>=1.2.71
brotli>=1.1.0
setuptools>=65.0.0
chardet>=5.0.0

# Optional dependencies for local development only
# These won't work on Heroku but are useful for local testing
# undetected-chromedriver>=3.5.0
# selenium>=4.0.0

# Existing dependencies
flask-cors

